Conflicts:
	basestation/src/blockWorld.cpp
	basestation/src/robot.cpp
	include/anki/cozmo/basestation/blockWorld.h
	include/anki/cozmo/basestation/robot.h

+++ coretech/common/include/anki/common/basestation/math/matrix.h
@@ -82,7 +82,7 @@ namespace Anki {
 
   // The type for representing matrix dimensions
   // (Should this be something else? unsigned int? uchar? int?)
-  using MatDimType = size_t;
+  using MatDimType = s32;
   
   // A class for small matrices, whose size is known at compile time
   template<MatDimType NROWS, MatDimType NCOLS, typename T>
